moved the top-level window implementation from AbstractVstEditorView to its separate component
