+++ src/palette/internal/paletteactionscontroller.cpp
@@ -25,7 +25,7 @@
 using namespace mu::palette;
 using namespace mu::ui;
 
-static const mu::UriQuery MASTER_PALETTE_URI("musescore://palette/masterpalette?sync=false");
+static const mu::UriQuery MASTER_PALETTE_URI("musescore://palette/masterpalette?sync=false&modal=false");
 static const mu::UriQuery SPECIAL_CHARACTERS_URI("musescore://palette/specialcharacters?sync=false");
 static const mu::UriQuery TIME_SIGNATURE_PROPERTIES_URI("musescore://palette/timesignatureproperties");
 static const mu::UriQuery EDIT_DRUMSET_URI("musescore://palette/editdrumset");
